requirements.txt
aiohttp==3.9.1
aiolimiter==1.2.1
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
asgiref==3.8.1
asyncio==3.4.3
attrs==25.1.0
beautifulsoup4==4.12.2
blinker==1.9.0
Brotli==1.1.0
cachetools==4.2.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsociety==3.2.9
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.25.2
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
logger==1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mutagen==1.47.0
numpy==2.2.3
openai==1.69.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
propcache==0.3.0
pyairtable==2.2.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2025.1
pywin32-ctypes==0.2.3
requests==2.31.0
rich==13.5.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.25
tornado==6.4.2
tqdm==4.66.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.0.5
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
yt-dlp==2025.2.19
